refactor: update gen scripts
