+++ .scripts/gen-config-dts.ts
@@ -1,0 +1,184 @@
+// The file is not designed to run directly. `cwd` should be project root.
+import fs from "fs";
+import path from "path";
+import process from "process";
+
+import { createChecker, PropertyMeta } from "vue-component-meta";
+
+import {
+    getComponents,
+    fileExist,
+    getFolders,
+    anonymiseTypes,
+} from "../packages/docs/src/utils.ts";
+
+const __dirname = process.cwd();
+
+const componentDirectory = "./packages/oruga/src/components";
+
+if (!fileExist(path.resolve(__dirname, componentDirectory)))
+    throw new Error("Path not exist: " + componentDirectory);
+
+console.log("Creating vue-component-meta checker...");
+// create component meta checker
+const checker = createChecker(
+    path.resolve(__dirname, "./packages/oruga/tsconfig.app.json"),
+    { printer: { newLine: 1 } },
+);
+
+// get all component folders
+const component_folders = getFolders(componentDirectory);
+
+function generateConfigType(filePath: string): void {
+    console.log(`Processing components...`);
+    const components = component_folders
+        .map((folder) => {
+            const name = folder.toLowerCase();
+            const folderPath = path.resolve(
+                __dirname,
+                componentDirectory,
+                folder,
+            );
+
+            // get all vue components in a component folder
+            const folderComponents = getComponents(folderPath);
+
+            // get all configurable props from all components in folder
+            const props = folderComponents
+                .flatMap((comp) => {
+                    const file = comp + ".vue";
+                    const componentPath = path.resolve(
+                        __dirname,
+                        componentDirectory,
+                        folder,
+                        file,
+                    );
+                    const meta = checker.getComponentMeta(componentPath);
+
+                    if (!meta.props.length) {
+                        console.warn(
+                            `Failure parsing component '${name}': no properties found.`,
+                        );
+                        console.log("Recreating vue-component-meta checker...");
+                    }
+
+                    return meta.props
+                        .filter((prop) => {
+                            // filter only class props and configurable props
+                            if (prop.name.includes("Class")) return true;
+                            if (prop.default?.includes("getDefault")) {
+                                const path = (prop.default.match(/"(.*?)"/) || [
+                                    "",
+                                ])[0];
+                                return path.includes(".");
+                            }
+                            return false;
+                        })
+                        .map((prop: PropertyMeta & { deprecated?: string }) => {
+                            // remove undefined because we wrap the object with partial
+                            if (prop.type.includes("| undefined"))
+                                prop.type = prop.type.replace(
+                                    " | undefined",
+                                    "",
+                                );
+
+                            // change type for class props
+                            if (prop.type === "ComponentClass")
+                                prop.type = "ClassDefinition";
+                            // override specific types
+                            else prop.type = anonymiseTypes(prop.type);
+
+                            if (prop.name.includes("Classes")) {
+                                prop.type = "Record<string, any>";
+                                return prop;
+                            }
+
+                            // change property name based on config path
+                            if (
+                                prop.default &&
+                                prop.default?.includes("getDefault")
+                            ) {
+                                let name = (prop.default.match(/"(.*?)"/) || [
+                                    "",
+                                ])[0];
+                                name = name.substring(1, name.length - 1);
+                                const split = name.split(".");
+                                name = split.length == 2 ? split[1] : split[0];
+                                if (prop.name !== name) prop.name = name;
+                            }
+
+                            // sort union types
+                            if (
+                                prop.type.includes(" | ") &&
+                                !prop.type.includes("(") &&
+                                !prop.type.includes("[")
+                            )
+                                prop.type = prop.type
+                                    .split(" | ")
+                                    .sort((a, b) => a.localeCompare(b))
+                                    .join(" | ");
+
+                            // set deprecated text if available
+                            prop.deprecated = prop.tags
+                                ?.filter((tag) => tag.name === "deprecated")
+                                .map((tag) => tag.text)
+                                .join("");
+
+                            return prop;
+                        });
+                })
+                // filter duplicates
+                .filter(
+                    (item, idx, self) =>
+                        idx === self.findIndex((p) => p.name === item.name),
+                );
+
+            return { name, props };
+        })
+        // sort components by name
+        .sort((a, b) => a.name.localeCompare(b.name));
+
+    console.log(`${components.length} components processed.`);
+
+    const code = `import type {
+    ClassDefinition,
+    ComponentConfigBase,
+    DynamicComponent,
+    Numberish,
+} from "@/types";
+
+// Auto generated component theme config declarations
+declare module "../index" {
+    interface OrugaConfig {
+        ${components.map(
+            ({ name, props }) =>
+                `${name.toLowerCase()}?: ComponentConfigBase &
+            Partial<{${props
+                .map(
+                    (prop) => `
+                /**
+                 * ${prop.description}${
+                     prop.deprecated
+                         ? `
+                 * @deprecated ${prop.deprecated}`
+                         : ""
+                 }
+                 */
+                ${prop.name}: ${prop.type};`,
+                )
+                .join("")}
+            }>;`,
+        ).join(`
+        `)}
+    }
+}
+`;
+
+    const file = path.resolve(__dirname, filePath);
+    fs.writeFileSync(file, code, { encoding: "utf8" });
+
+    console.log(`File '${file}' generated.`);
+}
+
+// generate component theme config types file
+generateConfigType("./packages/oruga/src/config.d.ts");
